+++ src/config.generated.ts
@@ -6,7 +6,7 @@
  *
  * To modify config, edit automatosx.config.json instead.
  *
- * Generated: 2025-10-27T04:29:04.518Z
+ * Generated: 2025-10-28T07:43:12.941Z
  */
 
 import type { AutomatosXConfig } from './types/config.js';
@@ -311,7 +311,7 @@ export const PRECOMPILED_CONFIG: AutomatosXConfig = {
  * Metadata about the precompiled config
  */
 export const PRECOMPILED_CONFIG_META = {
-  generatedAt: '2025-10-27T04:29:04.518Z',
+  generatedAt: '2025-10-28T07:43:12.941Z',
   sourceFile: 'automatosx.config.json',
   version: '5.6.28'
 } as const;